+++ grapple/types/interfaces.py
@@ -227,3 +227,25 @@ class StreamFieldInterface(graphene.Interface):
             return self.value.source
 
         return self.value
+
+
+def get_snippet_interface():
+    return import_string(grapple_settings.SNIPPET_INTERFACE)
+
+
+class SnippetInterface(graphene.Interface):
+    snippet_type = graphene.String(required=True)
+    content_type = graphene.String(required=True)
+
+    @classmethod
+    def resolve_type(cls, instance, info, **kwargs):
+        return registry.snippets[type(instance)]
+
+    def resolve_snippet_type(self, info, **kwargs):
+        return self.__class__.__name__
+
+    def resolve_content_type(self, info, **kwargs):
+        self.content_type = ContentType.objects.get_for_model(self)
+        return (
+            f"{self.content_type.app_label}.{self.content_type.model_class().__name__}"
+        )