+++ tests/settings_custom_interfaces.py
@@ -2,3 +2,4 @@ from settings import *  # noqa: F403
 
 
 GRAPPLE["PAGE_INTERFACE"] = "testapp.interfaces.CustomPageInterface"  # noqa: F405
+GRAPPLE["SNIPPET_INTERFACE"] = "testapp.interfaces.CustomSnippetInterface"  # noqa: F405